Start work on C version of text/richtext processor
The decision to move to C is mainly due to limitations of PL/pgSQL. The
work on the document layout processor revealed issues with navigating a
tree of JSON(b) objects, producing errors such as 'malformed array
literal: "(...)"' while assigning from a two-dimensional array into a
one-dimensional array of richtext AST nodes. Instead of trying to work
around these limitations (again), it has been decided to port the
current codebase to a different platform better suited for developing
such tools.

After some consideration, pure ANSI C89 has been chosen, while aiming
for compatibility with "traditional" compilers and making the
implementation fully cross-platform, therefore not event POSIX APIs are
being used. The goal here is to make this the last needed code porting
in this project, while enabling the embedding of this tool in many other
languages.

Other languages and C variants have been considered as well, but the
goal was to choose the most compatible option.