Add implementation of abstract layout resolver
The abstract layout resolver calculates the abstract layout a parsed
richtext document. An abstract layout is a sequence of layout blocks,
where each block is either a content block or a control block (page
break, same page content start/end). Each content block is composed of
paragraphs, either explicit or implicit, and each paragraph is composed
of explicit lines. Every line is represented by a sequence of segments,
which specify both line-specific layout and content formatting, and
carry the content composed of text and whitespace parsed ASTNodes.

Thus the abstract layout can be though of as a binary normalized
representation of a richtext document, meant for simplifying the
rendering or a richtext document to the output format (HTML, plaintext,
terminal escape sequences, etc).

The code is not tested yet, unit tests are yet to be provided.